pkg/config-v2: rename to pkg/config, and fix build
