+++ cmd/api/main.go
@@ -15,7 +15,7 @@
 package main
 
 import (
-	"openpitrix.io/openpitrix/pkg/config-v2"
+	"openpitrix.io/openpitrix/pkg/config"
 	"openpitrix.io/openpitrix/pkg/handlers/apps"
 )
