+++ pkg/config/default.go
@@ -5,23 +5,26 @@
 package config
 
 // DefaultConfigFile is the default config file.
-const DefaultConfigPath = "~/.openpitrix/config.yaml"
+const DefaultConfigPath = "~/.openpitrix/config.toml"
 
 // DefaultConfigContent is the default config file content.
 const DefaultConfigContent = `
 # OpenPitrix configuration
 # https://openpitrix.io/
 
-db_type: 'mysql'
-db_host: 'root:password@tcp(127.0.0.1:3306)/openpitrix'
-db_encoding: 'utf8'
-db_engine: 'InnoDB'
-
-host: '127.0.0.1'
-port: 8080
-protocol: 'http'
-uri: '/'
+Host = "0.0.0.0"
+Port = 8080
 
 # Valid log levels are "debug", "info", "warn", "error", and "fatal".
-log_level: 'warn'
+LogLevel = "warn"
+
+[Database]
+Type         = "mysql"
+Host         = "127.0.0.1"
+Port         = 3306
+Encoding     = "utf8"
+Engine       = "InnoDB"
+DbName       = "openpitrix"
+RootPassword = "password"
+
 `
