+++ pkg/handlers/apps/db_test.go
@@ -14,7 +14,7 @@ import (
 	"testing"
 	"time"
 
-	"openpitrix.io/openpitrix/pkg/config-v2"
+	"openpitrix.io/openpitrix/pkg/config"
 	"openpitrix.io/openpitrix/pkg/handlers/apps"
 )
